backend/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
sqlalchemy==2.0.41
pillow==11.2.1
pdf2image==1.17.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.4
protobuf==6.31.1
google-cloud-documentai==3.5.0
google-cloud-dlp==3.31.0
starlette==0.46.2
presidio_analyzer==2.2.358
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
